LeafScaner/requirements.txt
flask==3.1.3
flask-pymongo==3.0.1
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pillow==12.1.1
